refactor: simplify question module architecture and fix session-bound errors BREAKING CHANGE: remove DI container and persistence service, replace with simple async functions - Remove src/question/di/ directory and complex DI container system - Remove src/question/services/persistence_service.py causing DetachedInstanceError - Add src/question/service.py with simple async functions following quiz module pattern - Update question router to use async service functions with single-session contexts - Update quiz router, orchestrator, and dependencies to remove DI usage - Fix all session type annotations and SQL join queries
