--- backend/src/question/di/__init__.py
@@ -1,10 +1,0 @@
-"""Dependency injection module for question generation system."""
-
-from .container import DIContainer, ServiceLifetime, ServiceScope, get_container
-
-__all__ = [
-    "DIContainer",
-    "ServiceLifetime",
-    "ServiceScope",
-    "get_container",
-]
