+++ backend/src/question/services/__init__.py
@@ -2,10 +2,8 @@
 
 from .content_service import ContentProcessingService
 from .generation_service import GenerationOrchestrationService
-from .persistence_service import QuestionPersistenceService
 
 __all__ = [
     "ContentProcessingService",
     "GenerationOrchestrationService",
-    "QuestionPersistenceService",
 ]
